refactoring,fixing tiny issues
